requirements.txt
asgiref==3.7.2 
django>=2.2,<5.0
sqlparse==0.4.4 
tzdata==2023.3
djangorestframework==3.14.0 
pytz==2023.3
django-phonenumber-field==7.1.0
Pillow==10.0.0
django-countries==7.5.1
typing-extensions==4.7.1
django-randompinfield==0.4.7
python-decouple==3.8
 PyJWT==2.8.0 
 aiohttp==3.8.5 
 aiohttp-retry==2.8.3 
 aiosignal==1.3.1 
 async-timeout==4.0.3 
 attrs==23.1.0 
 certifi==2023.7.22 
 charset-normalizer==3.2.0 
 frozenlist==1.4.0  
 multidict==6.0.4 
 requests==2.31.0 
 twilio>=7,<8  
 yarl==1.9.2
 django-rest-auth==0.9.5 
six==1.16.0
cffi==1.15.1 
cryptography==41.0.3 
defusedxml==0.7.1 
django-allauth==0.54.0 
oauthlib==3.2.2 
pycparser==2.21 
python3-openid==3.2.0 
requests-oauthlib==1.3.
amqp==5.1.1
billiard==4.1.0 
celery==5.3.1 
click==8.1.7 
click-didyoumean==0.3.0 
click-plugins==1.1.1 
click-repl==0.3.0 
kombu==5.3.1 
vine==5.0.0
amqp==5.1.1
 billiard==4.1.0 
 celery==5.3.1 
 click==8.1.7 
 click-didyoumean==0.3.0 
 click-plugins==1.1.1 
 click-repl==0.3.0 
 colorama==0.4.6 
 kombu==5.3.1 
 prompt-toolkit==3.0.39 
 python-dateutil==2.8.2 
 vine==5.0.0 
 wcwidth==0.2.6
 django-extra-fields==3.0.2
 psycopg2==2.9.7
social-auth-app-django==5.3.0 
social-auth-core==4.4.2
cachetools==5.3.1 
google-auth==2.23.0 
pyasn1==0.5.0 
pyasn1-modules==0.3.0 
rsa==4.9 
urllib3>=1.26.0,<3.0
google-api-core==2.11.1 
google-api-python-client==2.100.0 
google-auth-httplib2==0.1.1 
googleapis-common-protos==1.60.0 
httplib2==0.22.0 
protobuf==4.24.3 
pyparsing==3.1.1 
uritemplate==4.1.1
deprecated==1.2.14 
django-oauth-toolkit==2.3.0 
drf_social_oauth2==2.1.3 
ecdsa==0.18.0 
jwcrypto==1.5.0 
python-jose==3.3.0 
wrapt==1.15.0
django-js-asset==2.1.0 
django-mptt==0.14.0
serpy==0.3.1
django-elasticsearch-dsl==7.3 
django-elasticsearch-dsl-drf==0.22.5 
django-nine==0.2.7 
elasticsearch==7.17.9 
elasticsearch-dsl==7.4.1 
django-haystack==3.2.1 
drf-haystack==1.8.12
django-filter==23.3
chardet==3.0.4 
googletrans==3.0.0 
h11==0.9.0 
h2==3.2.0 
hpack==3.0.0 
hstspreload==2023.1.1 
httpcore==0.9.1 
httpx==0.13.3 
hyperframe==5.2.0 
rfc3986==1.5.0 
sniffio==1.3.0
 basket==1.0
stripe==6.6.0
django-twilio==0.14.3.2
cachecontrol==0.13.1 
fcm-django==2.0.0 
firebase-admin==6.2.0 
google-cloud-core==2.3.3 
google-cloud-firestore==2.12.0 
google-cloud-storage==2.11.0 
google-crc32c==1.5.0 
google-resumable-media==2.6.0 
grpcio==1.58.0 
grpcio-status==1.58.0 
msgpack==1.0.6 
proto-plus==1.22.3
django-cors-headers==4.2.0
rest-social-auth==8.1.0 
social-auth-app-django==5.3.0 
social-auth-core==4.4.2
rest_framework_simplejwt===0.0.2
pyyaml==6.0.1
MarkupSafe-2.1.3 coreapi-2.3.3 coreschema-0.0.4 django-rest-swagger-2.2.0 itypes-1.2.0 jinja2-3.1.2 openapi-codec-1.3.2 simplejson-3.19.2